% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CFO_tables}
\alias{CFO_tables}
\title{Names of DIMA tables used for the BLM's Carlsbad Field Office}
\format{
## `CFO_tables`
A character vector with names of 17 DIMA tables.
}
\usage{
CFO_tables
}
\description{
Names of DIMA tables used for the BLM's Carlsbad Field Office
}
\keyword{datasets}
