% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DIMA_tablenames}
\alias{DIMA_tablenames}
\title{Names of all DIMA tables}
\format{
## `DIMA_tablenames`
A character vector with names of all 154 tables in DIMA 5.6.
}
\usage{
DIMA_tablenames
}
\description{
Names of all DIMA tables
}
\keyword{datasets}
