% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_CFO_tables.R
\name{extract_CFO_tables}
\alias{extract_CFO_tables}
\title{Extract DIMA tables for the BLM's Carlsbad Field Office}
\usage{
extract_CFO_tables(dima)
}
\arguments{
\item{dima}{A filepath to a DIMA database.}
}
\value{
A list of dataframes, each with the name and contents of a DIMA table.
}
\description{
Extract DIMA tables for the BLM's Carlsbad Field Office
}
