% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_DIMA.R
\name{extract_DIMA}
\alias{extract_DIMA}
\title{Extract DIMA tables}
\usage{
extract_DIMA(path, empty_tables = FALSE)
}
\arguments{
\item{path}{File path to a DIMA database.}

\item{empty_tables}{Should empty tables be returned?}
}
\value{
A named list of dataframes, each representing a table in DIMA.
}
\description{
`extract_DIMA()` extracts the contents of a DIMA database into a named list
of dataframes, optionally returning only non-empty DIMA tables.
}
