% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_dima_tables.R
\name{extract_dima_tables}
\alias{extract_dima_tables}
\title{Extract DIMA Database to R Dataframes}
\usage{
extract_dima_tables(dima_path, exclude_empty_tables = TRUE)
}
\arguments{
\item{dima_path}{A filepath to a DIMA database.}

\item{exclude_empty_tables}{Logical. If TRUE, tables with no data will not be returned.}
}
\value{
A list of dataframes, one for each DIMA table and with the same names.
}
\description{
`extract_dima_tables()` queries all rows of all tables for DIMA 5.6.

**Note: This function only works on Windows systems, due to unavailability
of a MS Access driver on other systems.**
}
