% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractDima2R-package.R
\docType{package}
\name{ExtractDima2R-package}
\alias{ExtractDima2R}
\alias{ExtractDima2R-package}
\title{ExtractDima2R: Extract DIMA Database to Dataframes}
\description{
A narrowly focused package for extracting data from a DIMA database to dataframes. This only works on Windows, due to unavailability of a MS Access driver on other systems.
}
\keyword{internal}
