% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_dima_tables.R
\name{prep_dima_tables}
\alias{prep_dima_tables}
\title{Prep Dima Tables}
\usage{
prep_dima_tables(
  path_dima,
  speciesstate,
  path_out = NULL,
  remove_invalid_keys = TRUE
)
}
\arguments{
\item{path_dima}{A filepath to a DIMA mdb.}

\item{speciesstate}{An arbitrary character string for project identification.}

\item{path_out}{A filepath to the directory where the outfile should be written.}

\item{remove_invalid_keys}{If TRUE, remove data that cannot be connected to a valid primary key.}
}
\value{
A list of dataframes.
}
\description{
`prep_dima_tables()` extracts DIMA tables and prepares them for terradactyl's
analysis functions. It is nearly identical to `terradactyl::dima_prep()`,
except that `prep_dima_tables()` adds the DIMA table tblPlotNotes to the output.
}
