% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_dima.R
\name{query_dima}
\alias{query_dima}
\title{Query a DIMA database}
\usage{
query_dima(dima_path, query)
}
\arguments{
\item{dima_path}{A filepath to a MS Access DIMA database.}

\item{query}{A list of named SQL queries.}
}
\value{
A list of query results.
}
\description{
Query a DIMA database
}
