% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_table_SQL.R
\name{query_table_SQL}
\alias{query_table_SQL}
\title{Construct a SQL query}
\usage{
query_table_SQL(table, cols = "*", condition = NULL)
}
\arguments{
\item{table}{The name of the table to query.}

\item{cols}{A character vector of column names to query.
Alternatively a character string with comma separated column names.}

\item{condition}{A logical condition to determine which rows to select.
Defaults to `NULL` to select all rows.}
}
\value{
A character string containing a valid SQL query for the contents of a table.
}
\description{
Construct a SQL query
}
\examples{
query_table_SQL("TableName", "Column_A")
query_table_SQL("TableName", c("ID", "Column_A"))
query_table_SQL("TableName", "Column_A, Column_B")
query_table_SQL("TableName", c("Column_A"), condition = "some condition")
}
