% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_dima.R
\name{extract_dima}
\alias{extract_dima}
\title{Extract DIMA Database to R Dataframes}
\usage{
extract_dima(dima_path)
}
\arguments{
\item{dima_path}{A filepath to a DIMA database.}
}
\value{
A list of dataframes, one for each DIMA table and with the same names.
}
\description{
`extract_dima()` queries all rows of all DIMA tables existing at the time
this function was created. Because there is no programatic way to retrieve
a list of all data table names from DIMA without privileged access, this
function cannot recognize if the DIMA schema has changed to add or delete
tables.

**Note: This function only works on Windows systems, due to unavailability
of a MS Access driver on other systems.**
}
